+++ CHANGELOG.md
@@ -51,6 +51,8 @@ speaking, considered to be the player's experience.
 - Made the build guidelines on the Standard tileset slightly brighter, to make
   them more visible.
 - Lined up the villager and peasant numbers on the Villages Report.
+- Increased the game speed of the unlimited speed by not drawing every frame,
+  but only 60 FPS.
 
 
 ## [3.2.2] — 2025-08-18